// ui/src/lib/address-builder/interface/address-form-data.interface.ts
export interface IAddressFormData {
    nombreCalleDireccion: string;
    numeroPrincipalDireccion: string;
    letraPrincipalDireccion: string;
    bisPrincipalDireccion: string;
    letraPrincipalDireccion2: string;
    referenciaCardinalDireccion: string;
    numeroSecundarioDireccion: string;
    letraSecundariaDireccion: string;
    numeroCasaDireccion: string;
    referenciaCardinalDireccionSecundario: string;
    complementoDireccion: string;
  }
  